+++ DesignerMacros/src/rdbms/key-automation-extensions/_common/constants.ts
@@ -1,0 +1,5 @@
+const primaryKeyStereotypeId = "b99aac21-9ca4-467f-a3a6-046255a9eed6";
+const relationalDatabaseId = "51a7bcf5-0eb9-4c9a-855e-3ead1048729c";
+const isManagedKey = "is-managed-key";
+const autoManageKeys = "auto-manage-keys";
+const isBeingDeletedByScript = "is-being-deleted-by-script";