+++ DesignerMacros/src/rdbms/key-automation-extensions/attribute-on-changed/attribute-on-changed.ts
@@ -2,7 +2,9 @@
 
 (async () => {
 
-if (application?.getSettings("ac0a788e-d8b3-4eea-b56d-538608f1ded9")?.getField("Key Creation Mode")?.value != "explicit" || element.getParent().getMetadata("auto-manage-keys") == "false"){
+if (application?.getSettings("ac0a788e-d8b3-4eea-b56d-538608f1ded9")?.getField("Key Creation Mode")?.value != "explicit" ||
+    element.getParent().getMetadata("auto-manage-keys") == "false"
+) {
     return;
 }